[IMP] website_sale: refactor / organize controllers; checkout: add validation for fields; add `website_featured` flag to product for giving precedence in listing; website_sale_event + website_sale_options: update controllers based on website_sale updates
