+++ addons/website_sale/controllers/__init__.py
@@ -1,1 +1,5 @@
-import main
+from . import main
+from . import shop
+from . import product
+from . import checkout
+from . import payment
