+++ addons/website_sale_delivery/controllers/main.py
@@ -2,10 +2,12 @@
 import openerp
 from openerp import http
 from openerp.http import request
-import openerp.addons.website_sale.controllers.main
+
+from openerp.addons.website_sale.controllers.payment import \
+    PaymentController as Base
 
 
-class website_sale(openerp.addons.website_sale.controllers.main.website_sale):
+class PaymentController(Base):
 
     @http.route(['/shop/payment'], type='http', auth="public", website=True)
     def payment(self, **post):
@@ -15,14 +17,18 @@ class website_sale(openerp.addons.website_sale.controllers.main.website_sale):
         if carrier_id:
             carrier_id = int(carrier_id)
         if order:
-            request.registry['sale.order']._check_carrier_quotation(cr, uid, order, force_carrier_id=carrier_id, context=context)
+            request.registry['sale.order']._check_carrier_quotation(
+                cr, uid, order, force_carrier_id=carrier_id, context=context)
             if carrier_id:
                 return request.redirect("/shop/payment")
 
-        res = super(website_sale, self).payment(**post)
+        res = super(PaymentController, self).payment(**post)
         return res
 
     def order_lines_2_google_api(self, order_lines):
         """ Transforms a list of order lines into a dict for google analytics """
-        order_lines_not_delivery = [line for line in order_lines if not line.is_delivery]
-        return super(website_sale, self).order_lines_2_google_api(order_lines_not_delivery)
+        order_lines_not_delivery = [
+            line for line in order_lines if not line.is_delivery
+        ]
+        return super(PaymentController, self).order_lines_2_google_api(
+            order_lines_not_delivery)
